% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pugmm.R
\name{plot.pugmm}
\alias{plot.pugmm}
\title{Plotting method for \code{pugmm} object}
\usage{
\method{plot}{pugmm}(x, ...)
}
\arguments{
\item{x}{Output from \code{pugmm}.}

\item{...}{Other graphics parameters.}
}
\description{
Plots for Parsimonious Ultrametric Gaussian Mixture Models results, such as BIC and path diagrams.
}
\details{
Types of graphs:
\enumerate{
\item BIC: plot of BIC values for the fitted models.
\item Path Diagram: path diagram representation of the component extended ultrametric covariance matrices for the best model.
}
}
\examples{
data(wine, package = "HDclassif")
x <- scale(wine[, -1])
pugmm.wine <- pugmm(x, 3, 5)
plot.pugmm(pugmm.wine)
}
\seealso{
\code{\link[=pugmm]{pugmm()}}
}
