% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions_PUGMM.R
\name{pugmm_available_models}
\alias{pugmm_available_models}
\title{PUGMM Model Names}
\usage{
pugmm_available_models()
}
\value{
Available models in PUGMM, i.e., the thirteen extended ultrametric covariance structures of PUGMM.
}
\description{
Description  of the model names used in the \emph{PUGMM} package.
}
\details{
The PUGMM model names in the \emph{PUGMM} package are characterized by four letters:
\itemize{
\item First letter: it refers to the variable-group membership matrix \eqn{V}, which can be equal (E) or free to vary (F) across components.
\item Second, third, fourth letters: they refer to the matrices of the group variances \eqn{\Sigma_{V}}, the within-group covariances \eqn{\Sigma_{W}} and the between-group covariances \eqn{\Sigma_{B}}, respectively, by indicating if they are unique (U, i.e., equal within and across components), isotropic (I, i.e., equal within components), equal (E, i.e., equal across components) or free to vary across components (F).
}
}
\examples{
pugmm_available_models()

}
\references{
Cavicchia, C., Vichi, M., Zaccaria, G. (2024) Parsimonious ultrametric Gaussian mixture models. \emph{Statistics and Computing}, 34, 108.
}
\seealso{
\code{\link[=pugmm]{pugmm()}}
}
