% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions_PUGMM.R
\name{rand.member}
\alias{rand.member}
\title{Random partition of objects into classes}
\usage{
rand.member(n.obs, G)
}
\arguments{
\item{n.obs}{Number of objects}

\item{G}{Number of classes}
}
\value{
A binary and row-stochastic matrix with \eqn{n.obs} rows and \eqn{G} columns.
}
\description{
Performs a random partition of objects into classes.
}
\details{
No empty classes can occur.
}
\examples{
rand.member(10, 3)
}
